Maps/Warnings: Reduce the amount of warnings/err...

+++ src/tools/ToolsCata/vmap_tools/vmap4_extractor/adtfile.cpp
@@ -214,7 +214,7 @@ bool ADTFile::init(uint32 map_num, uint32 tileX, uint32 tileY)
         }
 
         //======================
-        ADT.seek(nextpos);
+        ADT.seek(static_cast<int>(nextpos));
     }
 
     ADT.close();
